wip: poc for versioned cpt tests
