+++ qa/pom.xml
@@ -19,8 +19,6 @@
     <module>acceptance-tests</module>
     <module>archunit-tests</module>
     <module>util</module>
-    <module>versioned-tests</module>
-    <module>shared-tests</module>
   </modules>
 
   <build>
