--- qa/versioned-tests/pom.xml
@@ -1,36 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<project xmlns="http://maven.apache.org/POM/4.0.0"
-  xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
-  xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/xsd/maven-4.0.0.xsd">
-  <modelVersion>4.0.0</modelVersion>
-  <parent>
-    <groupId>io.camunda</groupId>
-    <artifactId>camunda-qa</artifactId>
-    <version>8.9.0-SNAPSHOT</version>
-  </parent>
-
-  <artifactId>versioned-tests</artifactId>
-  <packaging>jar</packaging>
-  <name>Camunda Versioned Tests</name>
-
-  <properties>
-    <!-- Allow overriding client version via -Dcamunda.client.version -->
-    <camunda.client.version>${project.version}</camunda.client.version>
-  </properties>
-
-  <dependencies>
-    <dependency>
-      <groupId>io.camunda</groupId>
-      <artifactId>camunda-client-java</artifactId>
-      <version>${camunda.client.version}</version>
-      <scope>test</scope>
-    </dependency>
-
-    <dependency>
-      <groupId>io.camunda</groupId>
-      <artifactId>shared-tests</artifactId>
-      <scope>test</scope>
-    </dependency>
-  </dependencies>
-
-</project>
