--- qa/versioned-tests/src/test/java/io/camunda/it/auth/AuthorizationSearchIT.java
@@ -1,15 +1,0 @@
-/*
- * Copyright Camunda Services GmbH and/or licensed to Camunda Services GmbH under
- * one or more contributor license agreements. See the NOTICE file distributed
- * with this work for additional information regarding copyright ownership.
- * Licensed under the Camunda License 1.0. You may not use this file
- * except in compliance with the Camunda License 1.0.
- */
-package io.camunda.it.auth;
-
-import io.camunda.qa.util.versioned.VersionedTest;
-import io.camunda.shared.auth.AuthorizationSearchITShared;
-
-@VersionedTest(enableAuthorization = true)
-public class AuthorizationSearchIT extends AuthorizationSearchITShared {
-}
