--- qa/versioned-tests/src/test/java/io/camunda/it/client/AuthorizationIntegrationTest.java
@@ -1,16 +1,0 @@
-/*
- * Copyright Camunda Services GmbH and/or licensed to Camunda Services GmbH under
- * one or more contributor license agreements. See the NOTICE file distributed
- * with this work for additional information regarding copyright ownership.
- * Licensed under the Camunda License 1.0. You may not use this file
- * except in compliance with the Camunda License 1.0.
- */
-package io.camunda.it.client;
-
-import io.camunda.qa.util.versioned.VersionedTest;
-import io.camunda.shared.client.AuthorizationIntegrationTestShared;
-
-@VersionedTest
-public class AuthorizationIntegrationTest extends AuthorizationIntegrationTestShared {
-
-}
